Fix when searching and only 1 disabled result
When searching, and the results contain exactly one result that is disabled,
it was previously "focused" (at least set as the `focusedOption` in state),
hence pressing enter caused it to be selected.

This fixes that, ensuring that:
1. A disabled option does not become the focusedOption
2. When there is no focusedOption, pressing enter does nothing

Closing the menu and leaving the selection empty would be another option,
but that would IMHO feel strange.

Fixes #378