Add support for assistive technologies
This adds support for aria. Some minor changes to the markup have been
made in order to get a better experience with a screen reader.

This should be regarded as a first step, and there will almost certainly
be more things we can do to improve this further, and fixes for other
screenreaders etc.