Update aria-only class to have absolute positioning
When the component is restyled using flex, the width of the aria-only box causes
the width of the `Select-multi-value-wrapper` to fluctuate when clicking into
the input, and starting to type. Making this `position: absolute` removes it from
the flow of flex, and thus solves this issue.

This seems to be a standard way to do this, while keeping it accessible. See:
http://a11yproject.com/posts/how-to-hide-content/
http://itstiredinhere.com/accessibility/