Clear this._callback on cache hit
Addresses the second issue mentioned in the linked issue. While the
previous commit makes this change unnecessary for regular use cases, the
issue is still present when debouncing. By clearing out
`this._callback`, intermediary searches (e.g. 'Ti' when 'Tim' is already
fetched) won't cause the options for 'Tim' to change when the request
for 'Ti' completes.