Refactor and expand end-to-end tests
This begins to use a more flexible approach of simplifying selectors
and using `.find` to locate specific elements within a particular
type of select component on the page.  This allows more flexibility,
and is easier to extend.

Additionally, the tests are now grouped into a few describe blocks,
based on the type of features being tested. So far, these are `Basic`,
`Grouped`, and `Clearable`.

Finally, some new tests are added for clearable single select components.