Update README.md with Typescript info
For the uninitiated, some possible points of confusion about Typescript and React Select are solved with these 2 short sentences.   There will be no need to edit them in the future... or, at least until there is no demand for v4.

BACKGROUND:
For the current HEAD of master, the version number in the `package.json` is v4.x  and all code is written in Typescript, although Typescript code will not be published until v5. 

A reasonable inference, if not the only one possible, is that v4 ships with the `.d.ts` files that reflect this Typescript code,  especially since the `package.json` for v4.2.3 indicates that they do.  

Additionally, absent clarification, a user might assume that the typings on the current master apply to v4.  In fact, they are incompatible with the @types version.